Fix Impossible Cast issues in MemoryTopicSub: - recoverSubscription() -- map is defined as LinkedHashMap<MessageId, Message> -- msg is defined as <map> entry.getValue() so must be a Message -- condition if (msg.getClass() == MessageId.class) could never be true -- no need to cast at all when using generics
- recoverNextMessages()
-- basically same code copy/pasted so same fix

Removed 2 conditions from ServerSessionPoolImpl that would result in impossible casts. Conditions removed were trying to cast ActiveMQQueueSession and ActiveMQTopicSession to ActiveMQSession which is illegal.

Since it isn't obvious what to do if you get an ActiveMQQueueSession or ActiveMQTopicSession from getServerSession() I make it fall back to the else condition which raises an async exception. This is better than getting a ClassCastException at runtime.

Remove impossible cast in MemoryMessageStore